// src/app/models/Episode.ts
export interface Episode {
  Type: string;
  Title: string;
  Year: number;
  Rated: string;
  Released: string;
  Season: number;
  Episode: number;
  Runtime: string;
  Genre: string;
  Director: string;
  Writer: string;
  Actors: string;
  Plot: string;
  Language: string;
  Country: string;
  Poster: string;
  Ratings: string[];
  imdbRating: string;
}